rename resource.rs -> extension.rs
