+++ src/hugr/rewrite/simple_replace.rs
@@ -209,17 +209,17 @@ mod test {
         BuildError, Container, DFGBuilder, Dataflow, DataflowHugr, DataflowSubContainer,
         HugrBuilder, ModuleBuilder,
     };
+    use crate::extension::prelude::USIZE_T;
     use crate::hugr::views::HugrView;
     use crate::hugr::{Hugr, Node};
     use crate::ops::OpTag;
     use crate::ops::{LeafOp, OpTrait, OpType};
-    use crate::resource::prelude::USIZE_T;
     use crate::types::{AbstractSignature, Type};
     use crate::{type_row, Port};
 
     use super::SimpleReplacement;
 
-    const QB: Type = crate::resource::prelude::QB_T;
+    const QB: Type = crate::extension::prelude::QB_T;
 
     /// Creates a hugr like the following:
     /// --   H   --
