+++ src/lib.rs
@@ -11,14 +11,14 @@
 
 pub mod algorithm;
 pub mod builder;
+pub mod extension;
 pub mod hugr;
 pub mod macros;
 pub mod ops;
-pub mod resource;
 pub mod std_extensions;
 pub mod types;
 mod utils;
 pub mod values;
 
+pub use crate::extension::Resource;
 pub use crate::hugr::{Direction, Hugr, HugrView, Node, Port, SimpleReplacement, Wire};
-pub use crate::resource::Resource;
