+++ src/ops/dataflow.rs
@@ -2,8 +2,8 @@
 
 use super::{impl_op_name, OpTag, OpTrait};
 
+use crate::extension::ResourceSet;
 use crate::ops::StaticTag;
-use crate::resource::ResourceSet;
 use crate::types::{AbstractSignature, EdgeKind, Type, TypeRow};
 
 pub(super) trait DataflowOpTrait {
