+++ src/std_extensions/arithmetic/int_ops.rs
@@ -4,11 +4,11 @@ use smol_str::SmolStr;
 
 use super::super::logic::bool_type;
 use super::int_types::{get_width, int_type};
-use crate::resource::prelude::ERROR_TYPE;
+use crate::extension::prelude::ERROR_TYPE;
 use crate::types::type_param::TypeParam;
 use crate::utils::collect_array;
 use crate::{
-    resource::{ResourceSet, SignatureError},
+    extension::{ResourceSet, SignatureError},
     types::{type_param::TypeArg, Type, TypeRow},
     Resource,
 };
